requirements.txt
# Bank Lending Strategy Optimizer - Python Dependencies
# Install with: pip install -r requirements.txt

# FRED API
fredapi>=0.5.0

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Database (PostgreSQL)
psycopg2-binary>=2.9.0
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional (for advanced features)
# statsmodels>=0.14.0  # For time series analysis (ARIMA)
# prophet>=1.1  # Facebook's forecasting library
# shap>=0.42.0  # For model explainability
